refactor config, add grpo ci
