+++ .github/workflows/e2e_gsm8k.yml
@@ -46,6 +46,10 @@ jobs:
         run: |
           ray stop --force
           bash tests/e2e/run_qwen_gsm8k_function_rm.sh
+      - name: Running gsm8k e2e training tests on 8 L20 GPUs with rmpad using function rm (GRPO)
+        run: |
+          ray stop --force
+          bash tests/e2e/run_qwen_gsm8k_function_rm_grpo.sh
       - name: Running gsm8k e2e without rmpad using function rm
         run: |
           ray stop --force
